requirements.txt
Flask==0.9
Flask-RESTful==0.2.1
GitPython==0.3.2.RC1
Jinja2==2.6
Werkzeug==0.8.3
async==0.6.1
gitdb==0.5.4
pycrypto==2.6
smmap==0.8.2
wsgiref==0.1.2
